requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
pandas==2.1.1
scikit-learn==1.3.1
numpy==1.24.3 